refactor(carousel): extract Slide component into its own file
- Moved `Slide` from `CarouselCreators.tsx` to a new file `CarouselCreatorsSlide.tsx` under `components/carousel/components/`.
- Renamed `Slide` to `CarouselCreatorsSlide` for better naming consistency.
- Updated imports in `CarouselCreators.tsx` to reference `CarouselCreatorsSlide` instead of inlined `Slide`.
